+++ crates/client/src/client/mod.rs
@@ -31,20 +31,3 @@ pub use self::client_connection::ClientConnection;
 #[allow(deprecated)]
 pub use self::client_future::{AsyncClient, ClientFuture, ClientHandle, ClientResponse};
 pub use self::memoize_client_handle::MemoizeClientHandle;
-
-/// This is an alias for [`trust_dns_proto::StreamHandle`]
-#[deprecated(note = "use [`trust_dns_proto::StreamHandle`] instead")]
-pub use proto::StreamHandle;
-
-/// This is an alias for [`trust_dns_proto::DnsStreamHandle`]
-#[deprecated(note = "use [`trust_dns_proto::DnsStreamHandle`] instead")]
-pub use proto::DnsStreamHandle as ClientStreamHandle;
-
-/// This is an alias for [`trust_dns_proto::RetryDnsHandle`]
-#[deprecated(note = "use [`trust_dns_proto::RetryDnsHandle`] instead")]
-pub use proto::RetryDnsHandle as RetryClientHandle;
-
-/// This is an alias for [`trust_dns_proto::SecureDnsHandle`]
-#[cfg(feature = "dnssec")]
-#[deprecated(note = "use [`trust_dns_proto::SecureDnsHandle`] instead")]
-pub use proto::SecureDnsHandle as SecureClientHandle;